Add files via upload
This is my first starter website for LMC2700! I wanted to start out with a potential layout for my personal page that we’ll be working on later. I got the color inspiration from a dress I bought recently and thought the colors looked cute together! I also wanted to try out using a GIF. I ran into some frustrations with uploading images, so I am going to keep practicing that. This was my first time using HTML or CSS and I found it to be a much simpler concept than I expected, but still have a lot I can work on.